+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added support for running multiple sessions in parallel using threading or multiprocessing
+- Each session has its own logger and log file
 
 ## [0.2.1] - 2025-10-29